+++ examples/szengine/find_network_by_record_id.py
@@ -3,10 +3,10 @@ from senzing import SzEngineFlags, SzError
 from senzing_core import SzAbstractFactoryCore
 
 build_out_degrees = 1
+build_out_max_entities = 10
 flags = SzEngineFlags.SZ_FIND_NETWORK_DEFAULT_FLAGS
 instance_name = "Example"
 max_degrees = 2
-max_entities = 10
 record_list = [("CUSTOMERS", "1001"), ("CUSTOMERS", "1009")]
 settings = {
     "PIPELINE": {
@@ -20,7 +20,9 @@ settings = {
 try:
     sz_abstract_factory = SzAbstractFactoryCore(instance_name, settings)
     sz_engine = sz_abstract_factory.create_engine()
-    result = sz_engine.find_network_by_record_id(record_list, max_degrees, build_out_degrees, max_entities, flags)
+    result = sz_engine.find_network_by_record_id(
+        record_list, max_degrees, build_out_degrees, build_out_max_entities, flags
+    )
     print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")