+++ examples/szengine/preprocess_record.py
@@ -5,6 +5,7 @@ from senzing import SzEngineFlags, SzError
 from senzing_core import SzAbstractFactoryCore
 
 instance_name = "Example"
+# TODO - Change when new default is in V4
 flags = SzEngineFlags.SZ_RECORD_DEFAULT_FLAGS
 record_definition = json.dumps(
     {